fix: migrate to layered structure
