+++ src/CartingService/BusinessLogic/Commands/RemoveCommand.cs
@@ -1,7 +1,7 @@
 ﻿using CartingService.BusinessLogic.Exceptions;
-using CartingService.BusinessLogic.Interfaces;
-using CartingService.Domain.Interfaces.Entities;
-using CartingService.Domain.Interfaces.Ports;
+using CartingService.BusinessLogic.Interfaces.Handlers;
+using CartingService.BusinessLogic.Interfaces.Ports;
+using CartingService.DataAccess.Interfaces;
 
 namespace CartingService.BusinessLogic.Commands;
