+++ src/CartingService/BusinessLogic/Interfaces/Handlers/ICommandHandler.cs
@@ -1,0 +1,6 @@
+namespace CartingService.BusinessLogic.Interfaces.Handlers;
+
+public interface ICommandHandler<T>
+{
+    Task Execute(T request, CancellationToken cancellationToken);
+}
