+++ src/CartingService/BusinessLogic/Interfaces/Handlers/IQueryHandler.cs
@@ -1,0 +1,6 @@
+namespace CartingService.BusinessLogic.Interfaces.Handlers;
+
+public interface IQueryHandler<TRequest, TResult>
+{
+    Task<TResult> Execute(TRequest request, CancellationToken cancellationToken);
+}
