--- src/CartingService/BusinessLogic/Interfaces/ICommandHandler.cs
@@ -1,6 +1,0 @@
-﻿namespace CartingService.BusinessLogic.Interfaces;
-
-public interface ICommandHandler<TRequest>
-{
-    Task Execute(TRequest request, CancellationToken cancellationToken);
-}
