--- src/CartingService/BusinessLogic/Interfaces/IQueryHandler.cs
@@ -1,6 +1,0 @@
-﻿namespace CartingService.BusinessLogic.Interfaces;
-
-public interface IQueryHandler<TRequest, TResponse>
-{
-    Task<TResponse> Execute(TRequest request, CancellationToken cancellationToken);
-}
