+++ src/CartingService/CartingService.sln
@@ -7,11 +7,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CartingService.BusinessLogi
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CartingService.DataAccess", "DataAccess\CartingService.DataAccess.csproj", "{410BDD59-20EA-4942-BF84-9049ACD8066E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CartingService.Domain", "Domain\CartingService.Domain.csproj", "{75787BC2-C45D-475A-B0AA-4802F58642A2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CartingService.Domain.UnitTests", "..\..\tests\UnitTests\Domain\CartingService.Domain.UnitTests.csproj", "{F65BB563-68EC-4A53-A51F-5A44E9211ACD}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CartingService.BusinessLogic.UnitTests", "..\..\tests\UnitTests\BusinessLogic\CartingService.BusinessLogic.UnitTests.csproj", "{B03C7CEB-11AC-4158-802F-451FFAFFA7B1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CartingService.DataAccess.UnitTests", "..\..\tests\UnitTests\DataAccess\CartingService.DataAccess.UnitTests.csproj", "{EE7F00B9-3B23-4A43-A2FD-3F4FE0809063}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -30,17 +28,13 @@ Global
 		{410BDD59-20EA-4942-BF84-9049ACD8066E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{410BDD59-20EA-4942-BF84-9049ACD8066E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{410BDD59-20EA-4942-BF84-9049ACD8066E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{75787BC2-C45D-475A-B0AA-4802F58642A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{75787BC2-C45D-475A-B0AA-4802F58642A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{75787BC2-C45D-475A-B0AA-4802F58642A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{75787BC2-C45D-475A-B0AA-4802F58642A2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F65BB563-68EC-4A53-A51F-5A44E9211ACD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F65BB563-68EC-4A53-A51F-5A44E9211ACD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F65BB563-68EC-4A53-A51F-5A44E9211ACD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F65BB563-68EC-4A53-A51F-5A44E9211ACD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B03C7CEB-11AC-4158-802F-451FFAFFA7B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B03C7CEB-11AC-4158-802F-451FFAFFA7B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B03C7CEB-11AC-4158-802F-451FFAFFA7B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B03C7CEB-11AC-4158-802F-451FFAFFA7B1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EE7F00B9-3B23-4A43-A2FD-3F4FE0809063}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE7F00B9-3B23-4A43-A2FD-3F4FE0809063}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE7F00B9-3B23-4A43-A2FD-3F4FE0809063}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE7F00B9-3B23-4A43-A2FD-3F4FE0809063}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
