--- src/CartingService/Domain/CartingService.Domain.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="Exceptions\" />
-    <Folder Include="Interfaces\Entities\" />
-    <Folder Include="Interfaces\Ports\" />
-  </ItemGroup>
-
-</Project>
